generic report functions moved to lib/report.ts
