+++ backend/src/app/routers/report.ts
@@ -4,6 +4,6 @@ import * as reportController from '../controllers/report';
 export const router: Router = Router({mergeParams: true});
 
 router.get('/project/hours', reportController.generateReportProjectByHours);
-router.get('/project/pounds', reportController.generateReportProjectByPounds);
+router.get('/project/cost', reportController.generateReportProjectByCost);
 router.get('/user/hours', reportController.generateReportUserByHours);
-router.get('/user/pounds', reportController.generateReportUserByPounds);
+router.get('/user/cost', reportController.generateReportUserByCost);
